Fix a duplicate warning in str-format-in-logging
Since this check works by walking up the parents of a function call, on
some versions of Python the logging functions (warning, error, debug,
etc.) have an instance type of `logging`. This means that when we
continue to walk back up to the logger object itself, we warn a second
time. This commit fixes this duplicate warning by breaking out of the
for loop once we've warned.